+++ src/problem.h
@@ -4,12 +4,12 @@
 struct {
      Point ORIGIN_POINT;
      int POINT_NUMBER;
-     int SPHERE_RADIUS;
-     int * (*satellite_distance_producer) (int);
-     void (*point_input_callback)(int, int *, int *, int *);
+     double SPHERE_RADIUS;
+     double * (*satellite_distance_producer) (int);
+     void (*point_input_callback)(int, double *, double *, double *);
      void (*logger)(char *);
      void (*format_logger) (char *, ...);
-     void (*result_callback)(int, int, int);
+     void (*result_callback)(double, double, double);
 } PROBLEM_INIT;
 
 void problem_solution();
